+++ Example/ViewController.swift
@@ -787,6 +787,30 @@ extension ViewController: NavigationMapViewDelegate {
 // MARK: - NavigationViewControllerDelegate methods
 
 extension ViewController: NavigationViewControllerDelegate {
+    
+    // To modify the width of the alternative route line layer through delegate methods.
+    func navigationViewController(_ navigationViewController: NavigationViewController, willAdd layer: Layer) -> Layer? {
+        guard var lineLayer = layer as? LineLayer else { return nil }
+        if lineLayer.id.contains("alternative.route_line") {
+            lineLayer.lineWidth = .expression(
+                Exp(.interpolate) {
+                    Exp(.linear)
+                    Exp(.zoom)
+                    RouteLineWidthByZoomLevel.multiplied(by: 0.7)
+                }
+            )
+        }
+        if lineLayer.id.contains("alternative.route_line_casing") {
+            lineLayer.lineWidth = .expression(
+                Exp(.interpolate) {
+                    Exp(.linear)
+                    Exp(.zoom)
+                    RouteLineWidthByZoomLevel
+                }
+            )
+        }
+        return lineLayer
+    }
 
     func navigationViewController(_ navigationViewController: NavigationViewController, didArriveAt waypoint: Waypoint) -> Bool {
         if let delegate = UIApplication.shared.delegate as? AppDelegate,