+++ Tests/MapboxNavigationTests/NavigationViewControllerTests.swift
@@ -547,6 +547,69 @@ class NavigationViewControllerTests: TestCase {
         XCTAssertFalse(style.layerExists(withId: layerIdentifier))
         XCTAssertFalse(style.sourceExists(withId: sourceIdentifier))
     }
+    
+    func testNavigationMapViewWillAddLayerDelegate() {
+        guard let dependencies = createDependencies() else { XCTFail("Dependencies are nil"); return }
+        
+        class NavigationViewControllerDelegateMock: NavigationViewControllerDelegate {
+            let expectedRouteLineOpacity: Double = 0.2
+            let expectedRouteCasingOpacity: Double = 0.3
+            let expectedRouteCasingWidth: Double = 10.0
+            let expectedRestrictedLineOpacity: Double = 0.4
+            
+            func navigationViewController(_ navigationViewController: NavigationViewController, willAdd layer: Layer) -> Layer? {
+                guard var lineLayer = layer as? LineLayer else { return nil }
+                if lineLayer.id.contains("main.route_line") {
+                    lineLayer.lineOpacity = .constant(expectedRouteLineOpacity)
+                }
+                if lineLayer.id.contains("main.route_line_casing") {
+                    lineLayer.lineOpacity = .constant(expectedRouteCasingOpacity)
+                    lineLayer.lineWidth = .constant(expectedRouteCasingWidth)
+                }
+                if lineLayer.id.contains("restricted_area_route_line") {
+                    lineLayer.lineOpacity = .constant(expectedRestrictedLineOpacity)
+                }
+                return lineLayer
+            }
+        }
+        
+        let delegateMock = NavigationViewControllerDelegateMock()
+        let navigationViewController = dependencies.navigationViewController
+        navigationViewController.delegate = delegateMock
+        navigationViewController.routeLineTracksTraversal = true // test the traversed route layer
+        _ = navigationViewController.view // trigger view load
+        
+        guard let style = navigationViewController.navigationMapView?.mapView.mapboxMap.style else {
+            XCTFail("Failed to get the MapView style object.")
+            return
+        }
+        
+        let route = dependencies.navigationService.route
+        let routeIdentifier = route.identifier(.route(isMainRoute: true))
+        let routeCasingIdentifier = route.identifier(.routeCasing(isMainRoute: true))
+        let restrictedIdentifier = route.identifier(.restrictedRouteAreaRoute)
+        let traversedIdentifier = route.identifier(.traversedRoute)
+        navigationViewController.navigationMapView?.showsRestrictedAreasOnRoute = true
+        navigationViewController.navigationMapView?.show([route])
+        
+        guard let routelineOpacity = style.layerPropertyValue(for: routeIdentifier, property: "line-opacity") as? Double,
+              let routeCasingOpacity = style.layerPropertyValue(for: routeCasingIdentifier, property: "line-opacity") as? Double,
+              let routeCasingWidth = style.layerPropertyValue(for: routeCasingIdentifier, property: "line-width") as? Double,
+              let restrictedOpacity = style.layerPropertyValue(for: restrictedIdentifier, property: "line-opacity") as? Double,
+              let traversedWidth = style.layerPropertyValue(for: traversedIdentifier, property: "line-width") as? Double else {
+            XCTFail("Route line layers should all be present.")
+            return
+        }
+        
+        XCTAssertEqual(delegateMock.expectedRouteLineOpacity, routelineOpacity, accuracy: 1e-3, "Failed to customize route line layer through delegate.")
+        XCTAssertEqual(delegateMock.expectedRouteCasingOpacity, routeCasingOpacity, accuracy: 1e-3, "Failed to customize route casing layer through delegate.")
+        XCTAssertEqual(delegateMock.expectedRouteCasingWidth, routeCasingWidth, accuracy: 1e-3, "Failed to customize route casing layer through delegate.")
+        XCTAssertEqual(delegateMock.expectedRestrictedLineOpacity, restrictedOpacity, accuracy: 1e-3, "Failed to customize route restricted area layer through delegate.")
+        XCTAssertEqual(delegateMock.expectedRouteCasingWidth, traversedWidth, accuracy: 1e-3,
+                       "The traversed route layer should have the same width as the main route casing layer.")
+        XCTAssertNil(style.layerPropertyValue(for: traversedIdentifier, property: "line-opacity") as? Double,
+                     "The traversed route layer shouldn't have other properties modified.")
+    }
 }
 
 extension NavigationViewControllerTests: NavigationViewControllerDelegate, StyleManagerDelegate {